url shortener. After review refactoring Autoproxing removed, interface creation is discussible
