+++ src/main/java/faang/school/urlshortenerservice/service/ExpiredHashCleanerService.java
@@ -1,0 +1,5 @@
+package faang.school.urlshortenerservice.service;
+
+public interface ExpiredHashCleanerService {
+    int cleanUpBatch();
+}
